# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountDeletionOut": {
        "properties": {
          "pending_storage_cleanups": {
            "minimum": 0.0,
            "title": "Pending Storage Cleanups",
            "type": "integer"
          },
          "status": {
            "enum": [
              "deleted",
              "cleanup_pending"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "pending_storage_cleanups"
        ],
        "title": "AccountDeletionOut",
        "type": "object"
      },
...
      "ConsentOut": {
        "properties": {
          "tracking_consent": {
            "enum": [
              "full",
              "minimal"
            ],
            "title": "Tracking Consent",
            "type": "string"
          }
        },
        "required": [
          "tracking_consent"
        ],
        "title": "ConsentOut",
        "type": "object"
      },
      "ConsentUpdate": {
        "properties": {
          "tracking_consent": {
            "pattern": "^(full|minimal)$",
            "title": "Tracking Consent",
            "type": "string"
          }
        },
        "required": [
          "tracking_consent"
        ],
        "title": "ConsentUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreferencesResetOut": {
        "properties": {
          "default_presentation_format": {
            "enum": [
              "detailed",
              "concise",
              "worked_example",
              "analogy"
            ],
            "title": "Default Presentation Format",
            "type": "string"
          },
          "presentation_affinity_rows_removed": {
            "title": "Presentation Affinity Rows Removed",
            "type": "integer"
          },
          "tutor_panel_open": {
            "title": "Tutor Panel Open",
            "type": "boolean"
          }
        },
        "required": [
          "default_presentation_format",
          "tutor_panel_open",
          "presentation_affinity_rows_removed"
        ],
        "title": "PreferencesResetOut",
        "type": "object"
      },
...
      "SettingsOut": {
        "properties": {
          "default_presentation_format": {
            "enum": [
              "detailed",
              "concise",
              "worked_example",
              "analogy"
            ],
            "title": "Default Presentation Format",
            "type": "string"
          },
          "tracking_consent": {
            "enum": [
              "full",
              "minimal"
            ],
            "title": "Tracking Consent",
            "type": "string"
          },
          "tutor_panel_open": {
            "title": "Tutor Panel Open",
            "type": "boolean"
          }
        },
        "required": [
          "tracking_consent",
          "default_presentation_format",
          "tutor_panel_open"
        ],
        "title": "SettingsOut",
        "type": "object"
      },
      "SettingsUpdate": {
        "properties": {
          "default_presentation_format": {
            "enum": [
              "detailed",
              "concise",
              "worked_example",
              "analogy"
            ],
            "title": "Default Presentation Format",
            "type": "string"
          },
          "tracking_consent": {
            "enum": [
              "full",
              "minimal"
            ],
            "title": "Tracking Consent",
            "type": "string"
          },
          "tutor_panel_open": {
            "title": "Tutor Panel Open",
            "type": "boolean"
          }
        },
        "required": [
          "tracking_consent",
          "default_presentation_format",
          "tutor_panel_open"
        ],
        "title": "SettingsUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me": {
      "delete": {
        "description": "Removes owned database records synchronously and records every private\nstorage target in the durable cleanup queue before removing its owner or\nsource row. The response reports pending file cleanup honestly.",
        "operationId": "delete_my_account_api_v1_me_delete",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete My Account",
        "tags": [
          "identity"
        ]
      },
      "get": {
        "description": "The authenticated learner. Identity comes from the trusted BFF headers,\nnever from anything the browser supplied directly.",
        "operationId": "read_me_api_v1_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Me",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/me/consent": {
      "patch": {
        "description": "Separated from core account settings on purpose: this is the one field\nthat gates behavioral telemetry (events/router.py), not anything the\ncore learning loop depends on -- declining it never breaks upload,\ngeneration, study, assessment, or recommendation.",
        "operationId": "update_consent_api_v1_me_consent_patch",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Consent",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/me/preferences/reset": {
      "post": {
        "operationId": "reset_preferences_api_v1_me_preferences_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Preferences",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/me/settings": {
      "get": {
        "operationId": "read_settings_api_v1_me_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Settings",
        "tags": [
          "identity"
        ]
      },
      "patch": {
        "operationId": "update_settings_api_v1_me_settings_patch",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}